fix the shelf

+++ parquet-column/src/main/java/org/apache/parquet/column/statistics/histogram/HistogramOptBuilder.java
@@ -52,7 +52,7 @@ public class HistogramOptBuilder {
             ColumnDescriptor columnDescriptor = messageType.getColumnDescription(new String[]{cols[i]});//?
 
             columnDescriptorMap.put(columnDescriptor,
-                    new HistogramOpts.HistogramEntry(Long.parseLong(minList[i]), Long.parseLong(maxList[i]), Integer.parseInt(countList[i])));
+                    new HistogramOpts.HistogramEntry((long) Double.parseDouble(minList[i]), (long) Double.parseDouble(maxList[i]), Integer.parseInt(countList[i])));
         }
 
         return new HistogramOpts(columnDescriptorMap);
